requirements.txt
affine
netCDF4
scipy
numpy
pyyaml
python-dateutil
